requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# CORS
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1